refactor: separate types from components
