+++ src/components/shared/Button/Button.types.ts
@@ -1,0 +1,11 @@
+export type ButtonVariant = "default" | "outline";
+
+/**
+ * Props for reuseable Button component
+ */
+export interface ButtonProps {
+  text: string;
+  variant?: ButtonVariant;
+  disabled?: boolean;
+  onClick: () => void;
+}
